+++ extension.js
@@ -32,6 +32,7 @@ var _DIR
 var currentRpcProvider
 var currentImlProvider
 var currentParametersProvider
+var currentStaticImlProvider
 var client
 
 async function activate(context) {
@@ -118,7 +119,7 @@ async function activate(context) {
         /**
          * Registering commands
          */
-        let coreCommands = new CoreCommands(appsProvider, _authorization, _environment, currentRpcProvider, currentImlProvider, currentParametersProvider)
+        let coreCommands = new CoreCommands(appsProvider, _authorization, _environment, currentRpcProvider, currentImlProvider, currentParametersProvider, currentStaticImlProvider)
         await CoreCommands.register(_DIR, _authorization, _environment)
         await AppCommands.register(appsProvider, _authorization, _environment, _admin)
         await ConnectionCommands.register(appsProvider, _authorization, _environment)